+++ src/web/aws-cloudwatch/StepHealthCheck.jsx
@@ -4,7 +4,7 @@ import { Col, Row } from 'react-bootstrap';
 
 import { Input } from 'components/bootstrap';
 
-import { LogOutputContext } from './providers/LogOutput';
+import { LogOutputContext } from './context/LogOutput';
 import FormWrap from '../common/FormWrap';
 
 const StepHealthCheck = ({ onSubmit }) => {
