+++ src/web/aws-cloudwatch/context/default_settings.js
